--- a/src/ext/java/org/opentripplanner/ext/flex/edgetype/FlexTripEdge.java
+++ b/src/ext/java/org/opentripplanner/ext/flex/edgetype/FlexTripEdge.java
@@ -91,4 +91,8 @@ public class FlexTripEdge extends Edge {
   public Trip getTrip() {
     return trip.getTrip();
   }
+
+  public FlexTrip getFlexTrip() {
+    return trip;
+  }
 }