--- a/src/ext/java/org/opentripplanner/ext/flex/trip/UnscheduledTrip.java
+++ b/src/ext/java/org/opentripplanner/ext/flex/trip/UnscheduledTrip.java
@@ -4,6 +4,7 @@ import org.opentripplanner.ext.flex.FlexServiceDate;
 import org.opentripplanner.ext.flex.flexpathcalculator.FlexPathCalculator;
 import org.opentripplanner.ext.flex.template.FlexAccessTemplate;
 import org.opentripplanner.ext.flex.template.FlexEgressTemplate;
+import org.opentripplanner.model.BookingInfo;
 import org.opentripplanner.model.FlexLocationGroup;
 import org.opentripplanner.model.StopLocation;
 import org.opentripplanner.model.StopTime;
@@ -34,6 +35,8 @@ public class UnscheduledTrip extends FlexTrip {
 
   private final UnscheduledStopTime[] stopTimes;
 
+  private final BookingInfo[] bookingInfos;
+
   public static boolean isUnscheduledTrip(List<StopTime> stopTimes) {
     Predicate<StopTime> noExplicitTimes = Predicate.not(st -> st.isArrivalTimeSet() || st.isDepartureTimeSet());
     Predicate<StopTime> notContinuousStop = stopTime ->
@@ -51,9 +54,11 @@ public class UnscheduledTrip extends FlexTrip {
     }
 
     this.stopTimes = new UnscheduledStopTime[N_STOPS];
+    this.bookingInfos = new BookingInfo[N_STOPS];
 
     for (int i = 0; i < N_STOPS; i++) {
       this.stopTimes[i] = new UnscheduledStopTime(stopTimes.get(i));
+      this.bookingInfos[i] = stopTimes.get(0).getBookingInfo();
     }
   }
 
@@ -127,6 +132,11 @@ public class UnscheduledTrip extends FlexTrip {
         .stream(stopTimes)
         .map(scheduledDeviatedStopTime -> scheduledDeviatedStopTime.stop)
         .collect(Collectors.toSet());
+  }
+
+  @Override
+  public BookingInfo getBookingInfo(int i) {
+    return bookingInfos[i];
   }
 
   private Collection<StopLocation> expandStops(StopLocation stop) {