--- a/src/ext/java/org/opentripplanner/ext/transmodelapi/model/plan/LegType.java
+++ b/src/ext/java/org/opentripplanner/ext/transmodelapi/model/plan/LegType.java
@@ -292,6 +292,7 @@ public class LegType {
             .newFieldDefinition()
             .name("bookingArrangements")
             .type(bookingArrangementType)
+            .dataFetcher(environment -> ((Leg) environment.getSource()).bookingInfo)
             .build())
         .build();
   }