--- a/src/main/java/org/opentripplanner/model/BookingInfo.java
+++ b/src/main/java/org/opentripplanner/model/BookingInfo.java
@@ -5,6 +5,12 @@ import org.opentripplanner.netex.mapping.BookingTime;
 import java.time.Duration;
 import java.util.EnumSet;
 
+/**
+ * Info about how a trip might be booked at a particular stop. All of this is pass-through
+ * information, except information about booking time and booking notice.
+ *
+ * // TODO Make the routing take into account booking time and booking notice.
+ */
 public class BookingInfo {
 
   private final ContactInfo contactInfo;