--- a/src/main/java/org/opentripplanner/model/BookingMethod.java
+++ b/src/main/java/org/opentripplanner/model/BookingMethod.java
@@ -1,0 +1,9 @@
+package org.opentripplanner.model;
+
+public enum BookingMethod {
+  CALL_DRIVER,
+  CALL_OFFICE,
+  ONLINE,
+  PHONE_AT_STOP,
+  TEXT_MESSAGE
+}