--- a/src/main/java/org/opentripplanner/model/ContactInfo.java
+++ b/src/main/java/org/opentripplanner/model/ContactInfo.java
@@ -1,6 +1,8 @@
 package org.opentripplanner.model;
 
-public class ContactInfo {
+import java.io.Serializable;
+
+public class ContactInfo implements Serializable {
 
   private final String contactPerson;