--- a/src/main/java/org/opentripplanner/netex/mapping/BookingMethodMapper.java
+++ b/src/main/java/org/opentripplanner/netex/mapping/BookingMethodMapper.java
@@ -1,0 +1,24 @@
+package org.opentripplanner.netex.mapping;
+
+import org.opentripplanner.model.BookingMethod;
+import org.rutebanken.netex.model.BookingMethodEnumeration;
+
+public class BookingMethodMapper {
+
+  public static BookingMethod map(BookingMethodEnumeration bookingMethodEnumeration) {
+    switch (bookingMethodEnumeration) {
+      case CALL_DRIVER:
+        return BookingMethod.CALL_DRIVER;
+      case CALL_OFFICE:
+        return BookingMethod.CALL_OFFICE;
+      case ONLINE:
+        return BookingMethod.ONLINE;
+      case PHONE_AT_STOP:
+        return BookingMethod.PHONE_AT_STOP;
+      case TEXT:
+        return BookingMethod.TEXT_MESSAGE;
+      default:
+        return null;
+    }
+  }
+}