… pom.xml
… src/main/java/org/opentripplanner/netex/mapping/BookingTime.java
… src/main/java/org/opentripplanner/netex/mapping/NetexMapper.java
… src/main/java/org/opentripplanner/netex/mapping/StopTimesMapper.java
… src/main/java/org/opentripplanner/netex/mapping/TripPatternMapper.java
… src/main/java/org/opentripplanner/routing/algorithm/mapping/RaptorPathToItineraryMapper.java
… src/main/java/org/opentripplanner/routing/trippattern/Deduplicator.java
… src/main/java/org/opentripplanner/routing/trippattern/TripTimes.java
… src/main/java/org/opentripplanner/standalone/OTPMain.java
… src/test/java/org/opentripplanner/netex/mapping/BookingInfoMapperTest.java
… src/test/java/org/opentripplanner/netex/mapping/StopTimesMapperTest.java
… src/test/java/org/opentripplanner/netex/mapping/TripPatternMapperTest.java
… src/test/java/org/opentripplanner/routing/trippattern/DeduplicatorTest.java