To be completed by pull request submitter:

- [ ] **issue**: #3343
- [ ] **roadmap**: Check the [roadmap](https://github.com/orgs/opentripplanner/projects/1) for this feature or bug. If it is not already on the roadmap, PLC will discuss as part of the review process.
- [ ] **tests**: Have you added relevant test coverage? Are all the tests passing on [the continuous integration service (Travis CI)](https://github.com/opentripplanner/OpenTripPlanner/blob/2.0-rc/docs/Developers-Guide.md#continuous-integration)?
- [ ] **formatting**: Have you followed the [suggested code style](https://github.com/opentripplanner/OpenTripPlanner/blob/2.0-rc/docs/Developers-Guide.md#code-style)? 
- [ ] **documentation**: If you are adding a new configuration option, have you added an explanation to the [configuration documentation](https://github.com/opentripplanner/OpenTripPlanner/blob/2.0-rc/docs/Configuration.md) tables and sections?
- [ ] **changelog**: add a bullet point to the [changelog file](https://github.com/opentripplanner/OpenTripPlanner/blob/2.0-rc/docs/Changelog.md) with description and link to the linked issue

To be completed by @opentripplanner/plc:

- [ ] reviews and approvals by 2 members, ideally from different organizations
- [ ] **after merging**: update the relevant card on the [roadmap](https://github.com/orgs/opentripplanner/projects/1)

This is the first step of implementing booking rules in OTP2. I have made the OTP model classes as well as the NeTEx import. This could then be integrated with the GTFS booking rules implementation from Kyyti (https://github.com/kyyticom/OpenTripPlanner/tree/flex2-booking-rules). Possibly some model changes will have to be made to accommodate this.

Tasks that are left:
- Unit tests for NeTEx import
- Implement GTFS import
- Add booking rules on legs. We should discuss how this should be done.
- Have OTP use booking notice and booking time values when routing.