@gmellemstrand 2021-02-12
@vpavlushkov @hannesj 

How does this suggestion work with what you have implemented in https://github.com/kyyticom/OpenTripPlanner/commits/flex2-booking-rules ?