@t2gran 2021-02-15
I tested the Deduplicator before and after my changes with the dataset for Norway(NeTEx) and San Fransisco Bay Area(GTFS). Below are a few log events for each test. Notice the differences between NeTEx and GTFS. The numbers for the Deduplicator are: `<unique instances>(<total including duplicates>)`.  

<details>
<summary>Testing the Deduplicator</summary>

```
###   SF BAY AREA   ###
OLD
Graph building took 2,9 minutes.
Main graph size: |V|=2231435 |E|=5755033
Save graph progress tracking complete. 449,5 MB done in 29s (15,3 MB pr second). 
Memory optimized Deduplicator{BitSet: 1,096, IntArray: 8,413, String: 80, StringArray: 143}
OTP SHUTTING DOWN (version: 2.1.0-SNAPSHOT, ser.ver.id: 2, commit: 21180e65

NEW 
Graph building took 2,8 minutes.
Main graph size: |V|=2231435 |E|=5755033
Save graph progress tracking complete. 449,7 MB done in 30s (14,8 MB pr second). 
Memory optimized Deduplicator{BitSet: 1096(117921), IntArray: 8413(0), String: 80(2107), StringArray: 143(0), BookingInfo: 101(0)}
OTP SHUTTING DOWN (version: 2.1.0-SNAPSHOT, ser.ver.id: 2, commit: 3d531893


####  NORWAY   ####

OLD
Graph building took 15,6 minutes.
Main graph size: |V|=2156808 |E|=5115131
Save graph progress tracking complete. 835,5 MB done in 36s (22,6 MB pr second). 
Memory optimized Deduplicator{BitSet: 1, IntArray: 30,899, String: 3,418, StringArray: 3,218}
OTP SHUTTING DOWN (version: 2.1.0-SNAPSHOT, ser.ver.id: 2, commit: 21180e65

NEW
Graph building took 15,0 minutes.
Main graph size: |V|=2156808 |E|=5115131
Save graph progress tracking complete. 836,1 MB done in 36s (22,7 MB pr second). 
Memory optimized Deduplicator{BitSet: 1(276966), IntArray: 30899(0), String: 3418(119301), StringArray: 3218(0), BookingInfo: 177(0)}
OTP SHUTTING DOWN (version: 2.1.0-SNAPSHOT, ser.ver.id: 2, commit: 3d531893

```

</details>